pyproject.toml
[project]
name = "rag-boticario"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.23",
    "langchain-chroma>=0.2.2",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.12",
    "langgraph>=0.3.29",
    "pydantic-settings>=2.8.1",
    "ruff>=0.11.5",
    "setuptools>=78.1.0",
]
